+++ src/game/offline/gameObjects/CentralTile.js
@@ -1,6 +1,5 @@
 import {STONE_TYPES} from "../../config.js";
 
-// TODO: write
 class CentralTile {
     constructor() {
         this.stonesTypes = [
@@ -10,11 +9,6 @@ class CentralTile {
             STONE_TYPES.GREEN,
             STONE_TYPES.GREEN,
             STONE_TYPES.BLUE];
-        this.stones = [];
-    }
-
-    getStone() {
-        return this.stones.shift();
     }
 }
