+++ src/game/offline/gameObjects/Player.js
@@ -2,7 +2,7 @@ export default class Player {
     constructor(id) {
         this.id = id;
         this.nickname = 'Player';
-        this.avatar = '';  // TODO: check usage i-gataullin
+        this.avatar = '';
         this.points = 0;
     }
 }
